+++ convert/src/coco/utils/render_pdf_page.py
