+++ convert/src/text_to_badgerdoc_converter.py
@@ -1,11 +1,11 @@
 import tempfile
 from pathlib import Path
 
-from src.badgerdoc_format.badgerdoc_format import BadgerdocFormat
-from src.badgerdoc_format.pdf_renderer import PDFRenderer
+from src.badgerdoc.badgerdoc import Badgerdoc
+from src.badgerdoc.pdf_renderer import PDFRenderer
 from src.config import DEFAULT_PAGE_BORDER_OFFSET
 from src.models.common import S3Path
-from src.plain_text_format.plain_text_converter import (
+from src.plain_text.plain_text_converter import (
     TextToBadgerdocTokensConverter,
 )
 
@@ -21,7 +21,7 @@ class TextToBadgerdocConverter:
         )
         self.pdf_renderer = PDFRenderer(page_border_offset=page_border_offset)
 
-        self.badgerdoc_format = BadgerdocFormat()
+        self.badgerdoc_format = Badgerdoc()
         self.s3_client = s3_client
 
     def execute(
